Created Chat store and fixed dense, center align comment
