+++ src/pages/PageAuth.vue
@@ -31,7 +31,7 @@
 <script>
 
 export default {
-    data () {
+  data () {
     return {
       tab: 'login'
     }
