+++ src/store/chat/getters.js
@@ -1,0 +1,11 @@
+export function users(state) {
+
+    let usersFiltered = {};
+    Object.keys(state.users).forEach((key) => {
+        if(key !== state.userDetails.userId) {
+            usersFiltered[key] = state.users[key];
+        }
+    })
+    return usersFiltered;
+}
+
