+++ src/store/chat/index.js
@@ -5,8 +5,8 @@ import * as actions from './actions'
 
 export default {
   namespaced: true,
+  state,
   getters,
   mutations,
-  actions,
-  state
+  actions
 }
