+++ src/store/chat/state.js
@@ -1,0 +1,7 @@
+export default function () {
+  return {
+    userDetails: {},
+    users: {}, 
+    messages: {}
+  }
+}
