+++ src/store/index.js
@@ -1,7 +1,9 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 
-import store from './store'
+import { firebaseAuth, firebaseDb } from 'boot/firebase';
+
+import store from './chat/index.js'
 
 Vue.use(Vuex)
