--- src/store/module-example/actions.js
@@ -1,2 +1,0 @@
-export function someAction (/* context */) {
-}
