--- src/store/module-example/getters.js
@@ -1,2 +1,0 @@
-export function someGetter (/* state */) {
-}
